\name{lmenssp-package}
\alias{lmenssp-package}
\alias{lmenssp-package}
\docType{package}
\title{
Linear Mixed Effects Models with Non-stationary Stochastic Processes
}
\description{
Obtains maximum likelihood estimates of the model parameters, filters, smooths 
and forecasts random components of the model for three non-stationary 
processes: 1) Brownian motion, 2) integrated Brownian motion, 3) integrated 
Ornstein-Uhlenbeck process
}
\details{
\tabular{ll}{
Package: \tab lmenssp\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-21\cr
License: \tab GPL (>=2) \cr
}
}
\references{
Diggle PJ (1988) An approach to the analysis of repeated measurements. \emph{Biometrics}, \bold{44}, 959-971.

Diggle PJ, Sousa I, Asar O (2014) Real time monitoring of progression towards renal failure in primary care patients.
Tentatively accepted by \emph{Biostatistics}.

Laird NM, Ware JH (1982) Random-effects models for longitudinal data. \emph{Biometrics}, \bold{38}, 134-147.

Ross SM (1996) \emph{Stochastic processes}. John Wiley & Sons, New Jersey.

Taylor JMG, Cumberland WG, Sy JP (1994) A Stochastic Model for Analysis of Longitudinal AIDS Data. 
\emph{Journal of the American Statistical Association}, \bold{89}, 727-736.
}

